refactor(weight): extract WeightProgress component and improve progress handling
- Moved weight progress UI and logic from WeightEvolution to a new WeightProgress component.
- Improved fallback handling and added warning for unhandled progress types.
- Updated imports and usage in WeightEvolution for better separation of concerns.
